Fix the MonoMac Template so it runs
The MonoMac project was missing a MainMenu item
which was causing the application to crash.
Has to add back the use of Nuget because MonoMac projects
to NOT use MSBuild/xbuild so it cannot use the new .props
system.

Updated the Nugets to download the 3.4.0 release. We should at
some point switch over to using the internal channel.